upgrade: [김원기] spring security+ jwt+ redis 기반 로그인 구현 [WMB-3]

+++ src/main/java/com/example/demo/security/utils/JwtUtil.java
@@ -31,7 +31,7 @@ public class JwtUtil {
         return false;
     }
 
-    public String generateToken(Map<String, Object> valueMap, int minutes) {
+    public String generateToken(Map<String, Object> valueMap, long minutes) {
         //헤더 부분
         Map<String, Object> headers = new HashMap<>();
         headers.put("typ", "JWT");
